// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  TRANSLATOR: 'TRANSLATOR',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const MangaStatus = {
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  HIATUS: 'HIATUS'
} as const

export type MangaStatus = (typeof MangaStatus)[keyof typeof MangaStatus]


export const Visibility = {
  PUBLIC: 'PUBLIC',
  UNLISTED: 'UNLISTED',
  PRIVATE: 'PRIVATE'
} as const

export type Visibility = (typeof Visibility)[keyof typeof Visibility]


export const CoinTransactionType = {
  TOPUP: 'TOPUP',
  PURCHASE: 'PURCHASE',
  REFUND: 'REFUND',
  ADJUSTMENT: 'ADJUSTMENT',
  BONUS: 'BONUS'
} as const

export type CoinTransactionType = (typeof CoinTransactionType)[keyof typeof CoinTransactionType]


export const CoinTransactionStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELED: 'CANCELED'
} as const

export type CoinTransactionStatus = (typeof CoinTransactionStatus)[keyof typeof CoinTransactionStatus]


export const TopupOrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  EXPIRED: 'EXPIRED',
  CANCELED: 'CANCELED',
  FAILED: 'FAILED'
} as const

export type TopupOrderStatus = (typeof TopupOrderStatus)[keyof typeof TopupOrderStatus]


export const TranslatorInviteStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
  CANCELED: 'CANCELED',
  REVOKED: 'REVOKED'
} as const

export type TranslatorInviteStatus = (typeof TranslatorInviteStatus)[keyof typeof TranslatorInviteStatus]